issue #81 Zombienoid. WIP. Actuators need to be reworked.

+++ src/ZBE_main/include/zombienoid/entities/adaptors/BlockConerAdaptor.h
@@ -35,7 +35,7 @@ public:
     delete s;
 
     zbe::AvatarEntityContainer<zbe::Avatar, zbe::Positionable<2>, zbe::Stated>* aeContainer (new zbe::AvatarEntityContainer<zbe::Avatar, zbe::Positionable<2>, zbe::Stated>(e, e, e));
-    zbe::AABB2D aabb({e->getX(), e->getY()},{e->getX()+e->getW(), e->getY()+e->getH()});
+    zbe::AABB2D aabb({(double)e->getX(), (double)e->getY()},{(double)e->getX()+e->getW(), (double)e->getY()+e->getH()});
     std::shared_ptr<zbe::StaticSolidAABB2D<R> > cObject(new zbe::StaticSolidAABB2D<R>(aabb));
     std::shared_ptr<zbe::VoidReactObject<R> > vro(new zbe::VoidReactObject<R>());
